requirements.txt
PyYAML==5.3
pytz==2021.1
numpy==1.18.1
pandas==0.25.3
statsmodels==0.11.1
fbprophet==0.6
pmdarima==1.5.3
tensorflow==2.1.0
xgboost==1.0.2
Flask==1.1.1
markdown==3.2.1
gunicorn==20.0.4
